#913 Move to_dict to the collection
