+++ parcels/particlesets/collectionsoa.py
@@ -1,3 +1,9 @@
+from datetime import timedelta as delta
+from operator import attrgetter
+from ctypes import Structure, POINTER
+from bisect import bisect_left
+from math import floor
+
 import numpy as np
 
 from .collections import ParticleCollection
@@ -7,10 +13,6 @@ from parcels.particle import ScipyParticle, JITParticle  # noqa
 from parcels.field import Field
 from parcels.tools.loggers import logger
 from parcels.tools.statuscodes import OperationCode
-from operator import attrgetter
-from ctypes import Structure, POINTER
-from bisect import bisect_left
-from math import floor
 
 try:
     from mpi4py import MPI
@@ -29,6 +31,25 @@ github relation: #913 (particleset_class_hierarchy)
 purpose: defines all the specific functions for a ParticleCollection, ParticleAccessor, ParticleSet etc. that relates
          to a structure-of-array (SoA) data arrangement.
 """
+
+
+def _to_write_particles(pd, time):
+    """We don't want to write a particle that is not started yet.
+    Particle will be written if particle.time is between time-dt/2 and time+dt (/2)
+    """
+    return (np.less_equal(time - np.abs(pd['dt']/2), pd['time'], where=np.isfinite(pd['time']))
+            & np.greater_equal(time + np.abs(pd['dt'] / 2), pd['time'], where=np.isfinite(pd['time']))
+            & (np.isfinite(pd['id']))
+            & (np.isfinite(pd['time'])))
+
+
+def _is_particle_started_yet(pd, time):
+    """We don't want to write a particle that is not started yet.
+    Particle will be written if:
+      * particle.time is equal to time argument of pfile.write()
+      * particle.time is before time (in case particle was deleted between previous export and current one)
+    """
+    return np.less_equal(pd['dt']*pd['time'], pd['dt']*time) | np.isclose(pd['time'], time)
 
 
 def convert_to_flat_array(var):
@@ -854,7 +875,7 @@ class ParticleCollectionSOA(ParticleCollection):
         cstruct = CParticles(*cdata)
         return cstruct
 
-    def toDictionary(self):     # formerly: ParticleSet.to_dict()
+    def toDictionary(self, pfile, time, deleted_only=False):
         """
         Convert all Particle data from one time step to a python dictionary.
         :param pfile: ParticleFile object requesting the conversion
@@ -863,10 +884,54 @@ class ParticleCollectionSOA(ParticleCollection):
         returns two dictionaries: one for all variables to be written each outputdt,
          and one for all variables to be written once
 
-         This function depends on the specific collection in question and thus needs to be specified in specific
-         derivatives classes.
+        This function depends on the specific collection in question and thus needs to be specified in specific
+        derivative classes.
         """
-        raise NotImplementedError
+
+        data_dict = {}
+        data_dict_once = {}
+
+        time = time.total_seconds() if isinstance(time, delta) else time
+
+        indices_to_write = []
+        if pfile.lasttime_written != time and \
+           (pfile.write_ondelete is False or deleted_only is not False):
+            if self._data['id'].size == 0:
+                logger.warning("ParticleSet is empty on writing as array at time %g" % time)
+            else:
+                if deleted_only is not False:
+                    if type(deleted_only) not in [list, np.ndarray] and deleted_only in [True, 1]:
+                        indices_to_write = np.where(np.isin(self._data['state'],
+                                                            [OperationCode.Delete]))[0]
+                    elif type(deleted_only) in [list, np.ndarray]:
+                        indices_to_write = deleted_only
+                else:
+                    indices_to_write = _to_write_particles(self._data, time)
+                if np.any(indices_to_write) > 0:
+                    for var in pfile.var_names:
+                        data_dict[var] = self._data[var][indices_to_write]
+                    pfile.maxid_written = np.maximum(pfile.maxid_written, np.max(data_dict['id']))
+
+                pset_errs = ((self._data['state'][indices_to_write] != OperationCode.Delete) & np.greater(np.abs(time - self._data['time'][indices_to_write]), 1e-3, where=np.isfinite(self._data['time'][indices_to_write])))
+                if np.count_nonzero(pset_errs) > 0:
+                    logger.warning_once('time argument in pfile.write() is {}, but particles have time {}'.format(time, self._data['time'][pset_errs]))
+
+                # ==== this function should probably move back somewhere into the particle-file instead of the to_dict ==== #
+                if time not in pfile.time_written:
+                    pfile.time_written.append(time)
+
+                if len(pfile.var_names_once) > 0:
+                    first_write = (_to_write_particles(self._data, time) & _is_particle_started_yet(self._data, time) & np.isin(self._data['id'], pfile.written_once, invert=True))
+                    if np.any(first_write):
+                        data_dict_once['id'] = np.array(self._data['id'][first_write]).astype(dtype=np.int64)
+                        for var in pfile.var_names_once:
+                            data_dict_once[var] = self._data[var][first_write]
+                        pfile.written_once.extend(np.array(self._data['id'][first_write]).astype(dtype=np.int64).tolist())
+
+            if deleted_only is False:
+                pfile.lasttime_written = time
+
+        return data_dict, data_dict_once
 
     def toArray(self):
         """
